+++ src/main/java/racingcar/StringParser.java
@@ -1,0 +1,15 @@
+package racingcar;
+
+import java.util.List;
+
+public class StringParser {
+    private static final String COMMA = ",";
+
+    public static List<Car> parseStringToList(List<Car> cars, String input) {
+        String[] carNames = input.split(COMMA);
+        for (String carName : carNames) {
+            cars.add(new Car(carName));
+        }
+        return cars;
+    }
+}
